move with_awaitable_senders into its own header
